create WalletProvider component
