+++ components/WalletProvider.tsx
@@ -1,0 +1,40 @@
+import '@fontsource/inter/300.css'
+import '@fontsource/inter/800.css'
+import '@rainbow-me/rainbowkit/styles.css'
+import {
+  getDefaultWallets,
+  RainbowKitProvider,
+  darkTheme,
+} from '@rainbow-me/rainbowkit'
+import { ReactNode } from 'react'
+import { chain, configureChains, createClient, WagmiConfig } from 'wagmi'
+import { publicProvider } from 'wagmi/providers/public'
+
+const { chains, provider } = configureChains(
+  [chain.mainnet, chain.polygon, chain.optimism, chain.arbitrum],
+  [publicProvider()],
+)
+
+const { connectors } = getDefaultWallets({
+  appName: 'Developer DAO School of Code',
+  chains,
+})
+
+const wagmiClient = createClient({
+  autoConnect: true,
+  connectors,
+  provider,
+})
+
+/**
+ * Context provider for RainbowKit wallet
+ */
+export const WalletProvider = ({ children }: { children: ReactNode }) => {
+  return (
+    <WagmiConfig client={wagmiClient}>
+      <RainbowKitProvider chains={chains} theme={darkTheme()}>
+        {children}
+      </RainbowKitProvider>
+    </WagmiConfig>
+  )
+}
